requirements.txt
# FastAPI Backend Dependencies
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2

# AI/ML Dependencies
transformers>=4.50.0
torch>=2.0.0
tokenizers>=0.22.0
datasets>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
openai>=1.0.0

# Additional utility dependencies
python-dateutil>=2.8.0
regex>=2023.0.0
safetensors>=0.4.0
